Move test_fetch_message_from_link to test_messages
