+++ tests/toolbox/test_messages.py
@@ -1,7 +1,23 @@
+from unittest import mock
+
 import hikari
 import pytest
 
 import toolbox
+
+
+@pytest.mark.asyncio
+async def test_fetch_message_from_link():
+    link = "https://discord.com/channels/574921006817476608/1010666418007719956/1012539497415704636"
+    channel_id = 1010666418007719956
+    message_id = 1012539497415704636
+
+    bot = mock.Mock()
+    bot.rest = mock.Mock()
+    bot.rest.fetch_message = mock.AsyncMock()
+
+    await toolbox.fetch_message_from_link(link, bot=bot)
+    bot.rest.fetch_message.assert_called_with(channel_id, message_id)
 
 
 def test_validate_embed_valid():
