+++ tests/toolbox/test_strings.py
@@ -1,5 +1,4 @@
 import datetime
-from unittest import mock
 
 import pytest
 
@@ -25,20 +24,6 @@ def test_is_invite_fullmatch():
 def test_is_invite_partial_match():
     assert toolbox.is_invite("https://discord.gg/Jx4cNGG", fullmatch=False)
     assert toolbox.is_invite("https://discord.gg/Jx4cNGG/RANDOM GARBAGE", fullmatch=False)
-
-
-@pytest.mark.asyncio
-async def test_fetch_message_from_link():
-    link = "https://discord.com/channels/574921006817476608/1010666418007719956/1012539497415704636"
-    channel_id = 1010666418007719956
-    message_id = 1012539497415704636
-
-    bot = mock.Mock()
-    bot.rest = mock.Mock()
-    bot.rest.fetch_message = mock.AsyncMock()
-
-    await toolbox.fetch_message_from_link(link, bot=bot)
-    bot.rest.fetch_message.assert_called_with(channel_id, message_id)
 
 
 def test_no_style():
